+++ Control/Joint/Effects.hs
@@ -2,6 +2,7 @@ module Control.Joint.Effects (module Exports) where
 
 import Control.Joint.Effects.Maybe as Exports
 import Control.Joint.Effects.Either as Exports
+import Control.Joint.Effects.Store as Exports
 import Control.Joint.Effects.State as Exports
 import Control.Joint.Effects.Writer as Exports
 import Control.Joint.Effects.Reader as Exports