+++ CHANGELOG.md
@@ -52,3 +52,4 @@
 * Rename `Adaptable` class to `Completable` and `adapt` to `complete`
 * Rename `Liftable` class to `Adaptable` and `lift` to `adapt`
 * Define `MonadTrans` instances for `TU` and `UT` schemes
+* Define `Store` datatype and define `Functor` and `Comonad` instances for it